+++ include/tao/json/internal/array_traits.hpp
@@ -49,21 +49,21 @@ namespace tao
             template< template< typename... > class Traits, typename Base >
             static bool equal( const basic_value< Traits, Base >& lhs, const T& rhs ) noexcept
             {
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_array() && ( p->unsafe_get_array().size() == rhs.size() ) && std::equal( rhs.begin(), rhs.end(), p->unsafe_get_array().begin() );
             }
 
             template< template< typename... > class Traits, typename Base >
             static bool less_than( const basic_value< Traits, Base >& lhs, const T& rhs ) noexcept
             {
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_array() ? std::lexicographical_compare( p->unsafe_get_array().begin(), p->unsafe_get_array().end(), rhs.begin(), rhs.end() ) : ( p->type() < type::ARRAY );
             }
 
             template< template< typename... > class Traits, typename Base >
             static bool greater_than( const basic_value< Traits, Base >& lhs, const T& rhs ) noexcept
             {
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_array() ? std::lexicographical_compare( rhs.begin(), rhs.end(), p->unsafe_get_array().begin(), p->unsafe_get_array().end() ) : ( p->type() > type::ARRAY );
             }
          };