+++ include/tao/json/internal/object_traits.hpp
@@ -54,7 +54,7 @@ namespace tao
                static const auto eq = []( const typename T::value_type& r, const std::pair< const std::string, basic_value< Traits, Base > >& l ) {
                   return ( l.first == r.first ) && ( l.second == r.second );
                };
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_object() && ( p->unsafe_get_object().size() == rhs.size() ) && std::equal( rhs.begin(), rhs.end(), p->unsafe_get_object().begin(), eq );
             }
 
@@ -70,14 +70,14 @@ namespace tao
             template< template< typename... > class Traits, typename Base >
             static bool less_than( const basic_value< Traits, Base >& lhs, const T& rhs ) noexcept
             {
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_object() ? std::lexicographical_compare( p->unsafe_get_object().begin(), p->unsafe_get_object().end(), rhs.begin(), rhs.end(), pair_less() ) : ( p->type() < type::OBJECT );
             }
 
             template< template< typename... > class Traits, typename Base >
             static bool greater_than( const basic_value< Traits, Base >& lhs, const T& rhs ) noexcept
             {
-               const auto* p = lhs.skip_raw_ptr();
+               const auto* p = lhs.skip_value_ptr();
                return p->is_object() ? std::lexicographical_compare( rhs.begin(), rhs.end(), p->unsafe_get_object().begin(), p->unsafe_get_object().end(), pair_less() ) : ( p->type() > type::OBJECT );
             }
          };